w4-analytics-engineering/requirements.txt
dbt-core
dbt-bigquery
Python >=3.11
DuckDB
Google-Cloud-SDK
dbt-bigquery